Updates to use self-signed naming convention, simplifies file structure
