--- rust/rebase/src/crosskey/mod.rs
@@ -1,1 +1,0 @@
-pub mod crosskey;
