+++ rust/rebase/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod crosskey;
 pub mod signer;
 pub mod schema;
 pub mod witness;
