+++ rust/rebase/src/schema/mod.rs
@@ -1,4 +1,4 @@
 pub mod basic_post;
 pub mod basic_profile;
-pub mod crosskey;
+pub mod self_signed;
 pub mod schema_type;
